.astype(int)

+++ docs/snippets/all/views/spatial2d.py
@@ -11,10 +11,10 @@ n = 150
 angle = np.linspace(0, 10 * np.pi, n)
 spiral_radius = np.linspace(0.0, 3.0, n) ** 2
 positions = np.column_stack((np.cos(angle) * spiral_radius, np.sin(angle) * spiral_radius))
-r = np.linspace(255, 255, n).astype(int)
-g = np.linspace(255, 0, n).astype(int)
-b = np.linspace(0, 255, n).astype(int)
-colors = np.dstack((r, g, b))[0]
+r = np.linspace(255, 255, n)
+g = np.linspace(255, 0, n)
+b = np.linspace(0, 255, n)
+colors = np.dstack((r, g, b))[0].astype(int)
 radii = np.linspace(0.01, 0.7, n)
 
 rr.log("points", rr.Points2D(positions, colors=colors, radii=radii))
